chore: button v1.1 - rename folder to components - added colorsyle properties - change Button.className and Button component accordingly
