+++ src/components/Button/Button.classnames.ts
@@ -1,6 +1,6 @@
 import { VariantProps, cva } from "class-variance-authority";
 
-export const ButtonCVA = cva("btn", {
+export const ButtonCVA = cva("text-slate-100 btn", {
   variants: {
     size: {
       xs: "btn-xs",
@@ -13,10 +13,20 @@ export const ButtonCVA = cva("btn", {
       outline: "btn-outline",
       ghost: "btn-ghost",
     },
+    colorStyle: {
+      primary: "btn-primary",
+      secondary: "btn-secondary",
+      accent: "btn-accent",
+      blue: "btn-info",
+      red: "btn-error",
+      yellow: "btn-warning",
+      green: "btn-green",
+    },
   },
   defaultVariants: {
     variant: "contain",
     size: "md",
+    colorStyle: "primary",
   },
 });
