+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/component";
+import { Button } from "@/components";
 
 export default function Home() {
   return (
